Se quitan comentarios innecesarios y se modifica el parámetro de unidad_ejecucion en la expedición de los contratos DVE
