+++ controllers/cambio_estado_contrato_valido.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/astaxie/beego"
 	"github.com/udistrital/administrativa_mid_api/models"
-	//. "github.com/mndrix/golog"
 	. "github.com/udistrital/golog"
 )
