+++ controllers/gestion_documento_resolucion.go
@@ -3,11 +3,8 @@ package controllers
 import (
 	"fmt"
 	"time"
-	//"strconv"
-	//"strings"
-	//"encoding/json"
+
 	"github.com/astaxie/beego"
-	//. "github.com/mndrix/golog"
 	"github.com/udistrital/administrativa_mid_api/models"
 )
 
@@ -61,7 +58,7 @@ func (c *GestionDocumentoResolucionController) GetContenidoResolucion() {
 	}
 
 	fecha_actual := time.Now().Format("2006-01-02")
-	query = "?query=DependenciaId:" + id_facultad +",FechaFin__gte:"+fecha_actual+",FechaInicio__lte:"+fecha_actual
+	query = "?query=DependenciaId:" + id_facultad + ",FechaFin__gte:" + fecha_actual + ",FechaInicio__lte:" + fecha_actual
 	if err := getJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudCore")+"/"+beego.AppConfig.String("NscrudCore")+"/jefe_dependencia/"+query, &jefe_dependencia); err == nil {
 		contenidoResolucion.OrdenadorGasto.NombreOrdenador = BuscarNombreProveedor(jefe_dependencia[0].TerceroId)
