+++ models/docente_planta.go
@@ -17,11 +17,3 @@ type DocentePlanta struct {
 	Id_carrera        string
 	Carrera           string
 }
-
-type ObjetoDocentePlanta struct {
-	DocenteCollection struct {
-		Docente []struct {
-			Planta                  string `json:"planta"`
-		} `json:"docentes"`
-	} `json:"docentesCollection"`
-}
