+++ models/objeto_docente_planta.go
@@ -1,0 +1,9 @@
+package models
+
+type ObjetoDocentePlanta struct {
+	DocenteCollection struct {
+		Docente []struct {
+			Planta string `json:"planta"`
+		} `json:"docentes"`
+	} `json:"docentesCollection"`
+}
